FPLA-3- review changes
