+++ service/src/main/java/uk/gov/hmcts/reform/fpl/service/DraftCMOService.java
@@ -9,16 +9,15 @@ import uk.gov.hmcts.reform.fpl.model.CaseData;
 import uk.gov.hmcts.reform.fpl.model.CaseManagementOrder;
 import uk.gov.hmcts.reform.fpl.model.HearingBooking;
 import uk.gov.hmcts.reform.fpl.model.common.Element;
-import uk.gov.hmcts.reform.fpl.model.common.dynamic.DynamicElementIndicator;
 import uk.gov.hmcts.reform.fpl.model.common.dynamic.DynamicList;
 import uk.gov.hmcts.reform.fpl.model.common.dynamic.DynamicListElement;
+import uk.gov.hmcts.reform.fpl.utils.HearingDateHelper;
 
 import java.time.LocalDate;
 import java.time.format.FormatStyle;
 import java.util.List;
 import java.util.Map;
 import java.util.Objects;
-import java.util.UUID;
 import java.util.stream.Collectors;
 
 import static org.apache.commons.lang3.StringUtils.isEmpty;
@@ -79,9 +78,9 @@ public class DraftCMOService {
     }
 
     public DynamicList buildDynamicListFromHearingDetails(List<Element<HearingBooking>> hearingDetails) {
-        List<HearingDate> hearingDates = hearingDetails
+        List<HearingDateHelper> hearingDates = hearingDetails
             .stream()
-            .map(element -> new HearingDate(element.getId(), element.getValue().getDate()))
+            .map(element -> new HearingDateHelper(element.getId(), element.getValue().getDate(), dateFormatterService))
             .collect(Collectors.toList());
 
         return DynamicList.toDynamicList(hearingDates, DynamicListElement.EMPTY);
@@ -89,22 +88,5 @@ public class DraftCMOService {
 
     public String convertDate(LocalDate date) {
         return dateFormatterService.formatLocalDateToString(date, FormatStyle.MEDIUM);
-    }
-
-    class HearingDate implements DynamicElementIndicator {
-
-        private LocalDate date;
-        private UUID id;
-
-        HearingDate(UUID id, LocalDate date) {
-            this.id = id;
-            this.date = date;
-        }
-
-        @Override
-        public DynamicListElement toDynamicElement() {
-            final String dateString = convertDate(date);
-            return DynamicListElement.builder().code(id.toString()).label(dateString).build();
-        }
     }
 }
